% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlas INE.R
\name{download_atlas}
\alias{download_atlas}
\title{Descarga del Atlas de Renta}
\usage{
download_atlas(filename)
}
\arguments{
\item{filename, }{hay que seleccionar que fichero queremos descargar de los 4 posibles.}
}
\value{
Fichero de csv con 4 posibilidades: Renta, Procedencia de la renta, Gini y Demografica
}
\description{
Esta funcion nos genera un csv en formato long de 4 posibles fichero del Atlas de Renta del INE: Renta, Procedencia, Gini, Demografia.
}
