% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Donwload_ECV.R
\name{Download_ecv}
\alias{Download_ecv}
\alias{download_ecv}
\title{Descarga de ECV}
\usage{
download_ecv(start = 16, end = 22)
}
\arguments{
\item{start}{Años desde el 8 hasta el numero 22}

\item{end}{Años desde el 8 hasta el numero 22}
}
\value{
Ficheros csv de los años seleccionados en una carpeta llamada ECV
}
\description{
Funcion de descarga de los ficheros anuales de la Encuesta de Condiciones de Vida
}
\examples{
download_ecv(start = 9, end = 22)
download_ecv(start = 9, end = 9)
}
