% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\docType{data}
\name{esadeecpol-palettes}
\alias{esadeecpol-palettes}
\alias{palettes_ecpol}
\alias{palette_ecpol}
\alias{palette_one}
\alias{palette_two}
\alias{palette_semaforoinverted}
\alias{palette_semaforo}
\alias{palette_redgreen}
\alias{palette_greenred}
\alias{palette_three}
\alias{palette_four}
\alias{gradient_palette_red}
\alias{gradient_palette}
\alias{gradient_palette_range}
\alias{print_palette}
\alias{print_gradient}
\title{Conjunto de Funciones}
\format{
An object of class \code{list} of length 8.

An object of class \code{palette} of length 1.

An object of class \code{palette} of length 2.

An object of class \code{palette} of length 3.

An object of class \code{palette} of length 3.

An object of class \code{palette} of length 2.

An object of class \code{palette} of length 2.

An object of class \code{palette} of length 3.

An object of class \code{palette} of length 4.

An object of class \code{matrix} (inherits from \code{array}) with 20 rows and 20 columns.
}
\usage{
palettes_ecpol

palette_ecpol(name, n)

palette_one

palette_two

palette_semaforoinverted

palette_semaforo

palette_redgreen

palette_greenred

palette_three

palette_four

gradient_palette_red(n)

gradient_palette(n)

gradient_palette_range(n)

print_palette(x, ...)

print_gradient
}
\description{
Estas funciones hacen cosas muy interesantes.
}
\keyword{datasets}
