% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonts.R
\name{fonts_ecpol}
\alias{fonts_ecpol}
\title{Fonts_ecpol nos importa todas las fuentes de esade para que no tengamos que importarlas #' de forma manual cada vez que hagamos graficos en R}
\usage{
fonts_ecpol()
}
\value{
Las fuentes de esade cargadas en el global environment de R
}
\description{
Fonts_ecpol nos importa todas las fuentes de esade para que no tengamos que importarlas #' de forma manual cada vez que hagamos graficos en R
}
