% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_colors.R
\docType{data}
\name{names}
\alias{names}
\title{Lista de colores EsadeEcPol para usar en ggploot}
\format{
An object of class \code{character} of length 11.
}
\usage{
names
}
\value{
vector con el nombre y el hexadecimal que representa
}
\description{
Esta función sirve para usar los colores en español en vez de en hexa
}
\examples{
colors()

}
\keyword{datasets}
