% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_colors.R
\name{printcolors}
\alias{printcolors}
\alias{print_colors}
\title{Lista de colores EsadeEcPol}
\usage{
print_colors()
}
\value{
Tabla con el nombre y el hexadecimal de los colores básicos de esade
}
\description{
Esta función nos devuelve una tabla con los nombres de los colores y su hexadecimal
}
\examples{
print_colors()
}
