% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_ecpol.R
\name{themes}
\alias{themes}
\alias{theme_ecpol}
\title{Tema de los graficos de EsadeEcPol,}
\usage{
theme_ecpol()
}
\value{
Tema creado por Jorge para visualizar los resultados
}
\description{
contiene todas las clases relevantes para la creacion de los graficos, además este fichero contiene tambien todas las paletas de colores que están predefinidas, tanto las discrerrtas como las continuas
}
