% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Directories.R
\name{directory_ecpol}
\alias{directory_ecpol}
\title{directory_ecpol genera, una vez en el directorio principal del paper, la estructura de carpetas con la que trabajamos y con esta funci칩n no tenemos que crear las carpetas de forma manual
Adem치s tambien genera el .gitignore para ahorrarnos ese trabajo.}
\usage{
directory_ecpol()
}
\value{
Las carpetas necesarias para nuestro trabajo
}
\description{
directory_ecpol genera, una vez en el directorio principal del paper, la estructura de carpetas con la que trabajamos y con esta funci칩n no tenemos que crear las carpetas de forma manual
Adem치s tambien genera el .gitignore para ahorrarnos ese trabajo.
}
