% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download EPA.R
\name{download_epa}
\alias{download_epa}
\title{Descarga de los microdatos de la EPA y unificacion de ficheros}
\usage{
download_epa(start_q = 1, end_q = 4, start_y = 5, end_y = 23)
}
\arguments{
\item{start_q}{Primer trimestre de descarga}

\item{end_q}{Ultimo trimestre de descarga}

\item{start_y}{Primer año de descarga}

\item{end_y}{Ultimo año de descarga}
}
\value{
Ficheros csv trimestrales en la carpeta EPA
}
\description{
La funcion descarga los ficheros de microdatos de la EPA desde el primer trimestre de 2005 hasta el ultimo trimestre existente
}
