% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_ecpol.R
\docType{data}
\name{esadeecpol-palettes}
\alias{esadeecpol-palettes}
\alias{palettes_ecpol}
\alias{palette_ecpol}
\alias{print_palette}
\alias{print_gradient}
\title{Funciones de paleta para esadeecpol}
\format{
An object of class \code{list} of length 8.
}
\usage{
palettes_ecpol
}
\description{
Descripción general de las funciones de paleta.
}
\details{
Este archivo documenta un conjunto de funciones de paleta para el paquete esadeecpol.
}
\keyword{datasets}
