% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_ecpol.R
\name{theme_ecpol}
\alias{theme_ecpol}
\title{Tema de los graficos de EsadeEcPol,}
\usage{
theme_ecpol()
}
\value{
Tema creado por Jorge para visualizar los resultados
}
\description{
Tema de los graficos de EsadeEcPol,
}
